+++ src/sdk/PnP.Core/Services/Core/ApiCall.cs
@@ -9,7 +9,7 @@ namespace PnP.Core.Services
     /// </summary>
     internal struct ApiCall
     {
-        internal ApiCall(string request, ApiType apiType, string jsonBody = null, string receivingProperty = null, bool loadPages = false, Dictionary<string, string> headers = null)
+        internal ApiCall(string request, ApiType apiType, string jsonBody = null, string receivingProperty = null, bool loadPages = false)
         {
             Type = apiType;
             Request = request;
@@ -29,10 +29,10 @@ namespace PnP.Core.Services
             LoadPages = loadPages;
             SkipCollectionClearing = false;
             ExecuteRequestApiCall = false;
-            Headers = headers;
+            Headers = null;
         }
 
-        internal ApiCall(List<Core.CSOM.Requests.IRequest<object>> csomRequests, string receivingProperty = null, Dictionary<string, string> headers=null)
+        internal ApiCall(List<Core.CSOM.Requests.IRequest<object>> csomRequests, string receivingProperty = null)
         {
             Request = null;
             Type = ApiType.CSOM;
@@ -52,7 +52,7 @@ namespace PnP.Core.Services
             LoadPages = false;
             SkipCollectionClearing = false;
             ExecuteRequestApiCall = false;
-            Headers = headers;
+            Headers = null;
         }
 
         /// <summary>
@@ -150,7 +150,7 @@ namespace PnP.Core.Services
         internal bool ExecuteRequestApiCall { get; set; }
 
         /// <summary>
-        /// Http-Headers for Request
+        /// Optional Http-Headers for Request to make
         /// </summary>
         internal Dictionary<string, string> Headers { get; set; } 
     }